fix: no image founde in delete asset helper

+++ src/utils/helper.util.js
@@ -2,6 +2,7 @@ const jwt = require('jsonwebtoken');
 const { z } = require('zod');
 const crypto = require('crypto');
 const fs = require('fs');
+const multer = require('multer');
 const config = require('../configs/general.config');
 
 function getOffset(listPerPage, currentPage = 1) {
@@ -127,14 +128,67 @@ function generateAssetUrl(fileName) {
 }
 
 function deleteAsset(path) {
-  if (fs.existsSync(path)) {
+  if (fs.existsSync(path) && !path.split('/').pop() === '') {
     fs.unlinkSync(path);
   }
+}
+
+function setStorage() {
+  const storage = multer.diskStorage({
+    destination: (req, file, cb) => {
+      cb(null, 'public/assets/images');
+    },
+    filename: (req, file, cb) => {
+      cb(null, `${Date.now()}-${file.originalname}`);
+    },
+  });
+
+  return storage;
+}
+
+function setFileFilter(
+  allowedTypes = ['image/jpg', 'image/jpeg', 'image/png'],
+  fileSize = 1024 * 1024 * 5,
+) {
+  return (req, file, cb) => {
+    if (!allowedTypes.includes(file.mimetype)) {
+      const error = new Error('Incorrect file');
+      error.code = 'INCORRECT_FILETYPE';
+      cb(error, false);
+    } else if (file.size > fileSize) {
+      const error = new Error('File size is too large');
+      error.code = 'FILE_TOO_LARGE';
+      cb(error, false);
+    }
+    cb(null, true);
+  };
+}
+
+/**
+ * @param {import('multer').Options} options
+ */
+
+function uploadFile(options, fieldName = 'image') {
+  const upload = multer(options).single(fieldName);
+
+  return (req, res, next) =>
+    upload(req, res, (err) => {
+      if (err) {
+        return errorResponse(res, err.message, null, 422);
+      }
+      if (!req.file) {
+        return errorResponse(res, `${fieldName} is required`, null, 400);
+      }
+      return next();
+    });
 }
 
 /* File End */
 
 module.exports = {
+  uploadFile,
+  setFileFilter,
+  setStorage,
   deleteAsset,
   getFilePath,
   generateAssetUrl,
